apps/engine/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pyspark==3.5.1
scikit-learn==1.8.0
pandas==2.3.3
pyarrow==23.0.1
pinecone==8.1.0
pinecone-text==0.11.0
sentence-transformers==5.2.3
groq==1.1.1
python-dotenv==1.2.2
python-multipart==0.0.22
pypdf==6.8.0
langchain-text-splitters==1.1.1
presidio_analyzer==2.2.360
presidio_anonymizer==2.2.360
python-pptx
